+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Add `processorResponseType` to `Transaction`, `AuthorizationAdjustment`, and `CreditCardVerification`.
 * Add `authorizationExpiresAt` to `Transaction`
 * Add support for additional PayPal options when vaulting a PayPal Order during customer update
+* Allow PayPal payment ID and payer ID to be passed during transaction create
 
 ## 3.36.0
 * Fix dispute results in transactions not showing the correct status sometimes